+++ src/dbg_watchs.h
@@ -15,8 +15,8 @@ namespace vscode
 
 	private:
 		void t_table(lua_State* L);
-		void t_set(lua_State* L, int n);
-		void t_get(lua_State* L, int n);
+		void t_set(lua_State* L, size_t n);
+		void t_get(lua_State* L, size_t n);
 
 	private:
 		size_t cur_;
